Add files via upload
Performed exploratory analysis on the data. Found missing values in the data and corrected them. For example, certain values were given as NA when in fact NA represents a factor level (i.e. NA == missing basement, not missing data). These were corrected by adding a new factor level called "None". Some numerical values which are a code for a type of building were read in as values rather than a factor. This was fixed by transforming that feature into a categorical variable. The remaining missing data consisted of only a handful of data points (<10) with missing data. These rows were dropped.

Issues: There are two issues which may influence the data analysis process. (1) For year of garage built, if a home does not have a garage the row values were set to 0. This will influence the results of linear models more than non-linear models. I recommend instead using the average/median home built age of the entire set (or using interpolation to estimate this value based on the other predictors). (2) For ratings given between 1 and 10, we assume that there is a linear increase between each of the ratings (i.e. a score of 4 to 5 is the same improvement as 9 to 10). In certain cases this may be relevant but I cannot tell just yet. If it becomes a problem this may have to be changed.