Add files via upload
Updated explore.R to produce two datasets. (1) It produces a dataset with 60 numerical/ordered categorical variables. (2) It produces 80 numerical/ordered categorical variables PLUS the other variables via one hot encoding. In this new dataset because of the decomposition of variables there are well over 200+ variables. This is experimental at least so both datasets are available.

Also, I removed the Id variable as it's not a predictor.

Added knn.R which includes analysis on the data as well as outliers using PCA analysis. The conclusion is the dimensionality reduction in this way is not very useful/effective since it still requires p>>5 for any useful amount of information to be extracted. I am currently in the process of working out another solution for this.

knn.R also includes two distance functions: (1) Minkowski Distance (generalized Euclidean distance) and the Mahalanobis Distance (which is very useful for data with high covariance). These functions will later be used in the hyperparameter tuning stage.